Extracted repeating code to a util.
