+++ packages/ckeditor5-package-tools/lib/utils/get-webpack-config-dll.js
@@ -13,8 +13,7 @@ const fs = require( 'fs' );
 const webpack = require( 'webpack' );
 const TerserPlugin = require( 'terser-webpack-plugin' );
 const CKEditorWebpackPlugin = require( '@ckeditor/ckeditor5-dev-webpack-plugin' );
-const { getPostCssConfig } = require( '@ckeditor/ckeditor5-dev-utils' ).styles;
-const getThemePath = require( './get-theme-path' );
+const commonWebpackConfig = require( './common-webpack-config' );
 
 module.exports = options => {
 	const packageJson = require( path.join( options.cwd, 'package.json' ) );
@@ -89,48 +88,6 @@ module.exports = options => {
 
 		plugins: webpackPlugins,
 
-		resolve: {
-			// Add support for TypeScript files and fallback to default extensions list.
-			extensions: [ '.ts', '...' ]
-		},
-
-		module: {
-			rules: [
-				{
-					test: /\.svg$/,
-					use: 'raw-loader'
-				},
-				{
-					test: /\.css$/,
-					use: [
-						{
-							loader: 'style-loader',
-							options: {
-								injectType: 'singletonStyleTag',
-								attributes: {
-									'data-cke': true
-								}
-							}
-						},
-						'css-loader',
-						{
-							loader: 'postcss-loader',
-							options: {
-								postcssOptions: getPostCssConfig( {
-									themeImporter: {
-										themePath: getThemePath( options.cwd )
-									},
-									minify: true
-								} )
-							}
-						}
-					]
-				},
-				{
-					test: /\.ts$/,
-					use: 'ts-loader'
-				}
-			]
-		}
+		...commonWebpackConfig( options.cwd )
 	};
 };
